+++ test/jest/printHelp.test.js
@@ -1,5 +1,5 @@
 const { printHelp, loggerUtil } = require('../../output/')
-const { decolor }  = loggerUtil
+const { decolor } = loggerUtil
 
 test('printHelp(op)', () => {
   const checkMap = [
@@ -28,6 +28,50 @@ test('printHelp(op)', () => {
         '  Options:',
         '    -v, --version  show version',
         '    -p, --path     path to plugin',
+        '',
+        ''
+      ]
+    },
+    {
+      op: {
+        silent: true,
+        usage: 'yyl-print',
+        desc: '打印命令',
+        commands: {
+          init: 'init files <Package>',
+          default: 'default system'
+        },
+        options: {
+          '-v, --version': 'show version',
+          '-p, --path': 'path to plugin'
+        },
+        others: {
+          Package: {
+            packageA: '项目A',
+            packageB: '项目B'
+          }
+        }
+      },
+      result: [
+        '',
+        '  Usage: yyl-print <commands> <options>',
+        '',
+        '  打印命令',
+        '',
+        '',
+        '  Commands:',
+        '    init     init files <Package>',
+        '    default  default system',
+        '',
+        '',
+        '  Options:',
+        '    -v, --version  show version',
+        '    -p, --path     path to plugin',
+        '',
+        '',
+        '  Package:',
+        '    packageA  项目A',
+        '    packageB  项目B',
         '',
         ''
       ]